TiffRGBAImage: Return arbitrary image sections
Enable TIFFRGBAImage to work on arbitrary areas of the image. Previously, the
starting column had to be on a tile boundery. Now, all starting points and
sizes are supported.

Changes necessary:
 * Change the gtContigTile() get function to deal with col_start not being on a
   tile boundary.
 * Fix the 16 bit black/white put function to correctly deal with skew in the
   b/w source.
 * Fix the custom 16 bit b/w put function in a similar fashion.

This allows not only reading 512x512 tiles for map display, but also small
scale access to individual pixels or small areas, e.g. under the mouse
pointer to calculate terrain height and steepness.