Add a gradient map type for DHM coloration
Add a RasterMap::TYPE_GRADIENT map type. Currently that does some processing on
a DHM map and creates a colorized hight map. The resulting maps are actually
rather pretty and show hights and slopes very well.

Several infrastructure updates are also included, which allow one map (a DHM)
to be included in the RasterMapCollection multiple times with different display
types.