Improve map change and MapDisplay repaint logic
First, determining which map is currently active was moved into class
MapDisplayManager and away from MapCollection. It really is a display property,
not one of the underlaying collection.

Second, map change now optionally keeps the map center the same, if it makes
sense (i.e. the point shown on the old map is also on the new one).

Also make MapDisplayManager take care of repainting the map panel itself. This
allows for simplification of control flow all around and RootWindow is less in
the center of the whole application.