New abstraction layer for map modes
Introduce class UIMode and make the relevant UI interaction go through that
layer. This should allow us to introduce other modes later on (e.g. distance
measurement).

Also fix the mouse message handling:
We were querying the current mouse position instead of relying on the values in
the window messages. This gave weired behavior if ODM was lagging and the mouse
moved between message-send-time and message-processing.