Separate map coords from base map coords
* MapPixelCoord and friends describe map pixels within any given map.
* BaseMapCoord is a special MapPixelCoord that describes coordinates within the
  current active base map.

This distinction becomes important when we overlay several maps in the same map
panel. The base map dictates the pixel grid along with projection, direction of
north, etc. Other maps will typically be distorted compared to the base map.

Conversion from DisplayCoord to BaseMapCoords is trivial (via
MapDisplayManager.m_center and .m_zoom). Conversion of display coordinates to
arbitrary MapPixelCoords is trickier and we need to go via the base map and
Lat/Lon coordinates.