Improve Bezier code
Rework the code to sample directly from the source data, without first
calculating the Bezier coefficients separately. Also get rid of class MapBezier
and use a codelayout that makes it easy to use correctly and fast (without
either dangling pointers nor useless copying of data).

Also add a function to calculate the gradient at 0.5/0.5 (that is, directly on
top of a pixel) quickly, as that's a huge shortcut. Use that in
map_dhm_advanced. The time required to render a tile dropped from ~270 ms to
75 ms.