Improve build architecture
Rearranged things a bit so a central ./build.sh can build the whole project:
  $ git clone $GIT_REPO .
  $ ./build.sh DEBUG

Both DEBUG and RELEASE work, changing between them should also work, but requires redownloading and rebuilding all the libraries.

It might be nice to actually have 2 library trees, one for the release build
and one for the debug build, with the project/solution configured to
automatically pick the right one.