Support model=Geographic maps; Fix error handling
Add support for Tiff Model == Geographic maps, i.e. maps where x/y coordinates
in the image map directly to latitude and longitude. With these maps a
projection step is not really necessary as the transformation Projected CS <->
LatLong is a trivial lat = y; long = x; mapping.

In the course of these changes it also became necessary to fix the error
handling so that maps without valid projection work well. This is actually the
larger part of the patch, but was necessary to get model-Geographic maps
working.