Add crude map overlay functionality
Add crude map overlays:
To overlay one map over the other, first load the base map (the lowest one).
Then right-click on the map-to-overlay in the map database.

Currently, multiple overlay maps are supported, but not very useful.
Transparency of the overlaid maps is fixed to 50%.

To get rid of the overlay, change the base map.