Make map data (unsigned int*) a first-class object
Introduce class MapRegion that encapsulates plain map data, internally stored
as shared_ptr<unsigned int>. Return MapRegion from RasterMap::GetRegion() and
convert all the implementing and calling code.

This change is inspired for the need for some coherent python interface later
on, which is more difficult to do if we work with shared_ptr's.