Fix shared-lib vs executable defines, add DllMain
Add a new optional #define ODM_EXECUTABLE, which indicates that an executable
is being built. It is mutually exclusive with ODM_SHARED_LIB. This is required
to set EXPORT to the correct value for each case.

Further add a DllMain function and save our DLL's HMODULE. Do not include
WinMain in the DLL builds any more.