Make use of shared_ptr's more consistent
Previously, we used:
* shared_ptr<const RasterMap>
* shared_ptr<RasterMap>
* RasterMap*
* RasterMap&

While not elegant in C++, it worked OK in practice. With SIP it becomes a
problem, as the transformation between the different shared_ptr's and the base
pointers isn't so easy there.

Now we consistently pass RasterMaps around as shared_ptr<RasterMap>.
const/non-const doesn't matter much, since class RasterMap is read-only anyway
in practice. While fixing the rippling effects of that change, also improve
coding style and a few other issues around the code base.