Make SmartptrProxy comparison work properly
SmartptrProxy object comparison was broken due to the fact that the default
tp_richcompare simply compared the SmartptrProxyObject* pointers. For example,
comparing two Rastermaps yielded this:

    >> rastermapcollection.Get(0) == rastermapcollection.Get(0)
    False

__getattr__ didn't help, because it resulted in a call like
``target.__class__.tp_richcompare(proxy1, proxy2)``, where the target
richcompare implementation was invoked on the proxy objects. If the target
didn't have a custom richcompare handler, we fell back to the default pointer
comparison.

We solve this issue by providing our own tp_richcompare handler, which checks
if both objects are instances of SmartptrProxy. If no, we let the default
handler do it's thing (via a base class tp_richcompare call). If yes, we return
the rich comparison of the target objects.