Let maps draw more directly to the screen.
The old tile logic was problematic for overlayed data, such as GPS tracks:

* There is no "correct" resolution for GPS tracks. When zoomed out, they became
  nearly invisible, and when zoomed in there was no way to provide better
  resolution or more information for the user.
* The tiles were post-processed in the display layer by rotating/stretching
  them to fit the base map projection. This caused bluring and reduced display
  quality.

So we introduce a new drawing mode, termed "Direct", for lack of a better name.
In it, the map is supplied with a reference to the base map and the display
size, and produces an image that exactly fits the display. Said image is then
directly drawn without further manipulation. Here, the map itself is
responsible for mapping display pixels to LatLon, and can decide where it wants
to draw.

Supporting this throughout the drawing pipeline took a number of changes:
 * class RasterMap gained a SupportsDirectDrawing() and a GetRegionDirect()
   method. Currently, only class GPSSegment implements that new interface.
 * class DisplayOrder is now an abstract base class, with 2 implementations:
   DisplayOrderTiled and DisplayOrderDirect.
 * class MapDisplayManager was restructured to better encapsulate the old tile
   based logic. Additionally, DisplayOrderDirect is now used when appropriate.
 * class DispOpenGL now gets its image data directly from class DisplayOrder,
   instead of relying on class TileCode. The latter serves only as tag for
   caching tiles. That makes DispOpenGL independent of the DisplayOrder
   implementation used.
   Note: DispOpenGL doesn't care too much about the new-fangled Direct stuff.
   All it sees is an oddly-sized, non-cacheable tile it must draw. It just so
   happens to fill the entire output area.