Add support for drawables with alpha channel
Add support for different pixel formats supported by maps. Currently, this is
used to support an alpha channel for GPS tracks, so only the track points are
drawn and the rest of the overlay is transparent.

Eventually, it may be possible to visualize DHMs directly via glTexImage1D().
Further, other/more pixel formats may be supported (e.g. 24 bit RGB).

Currently, the transparency slider is broken for maps specifying their own
alpha channel. This is due to OpenGL blending either with the texture alpha OR
a constant alpha, but not both. This may be rectified using multitexturing.