Add MapsEvolved.bat for running on Windows
tools/MapsEvolved.bat kills previously running instances, loads the venv and
executes MapsEvolved. After the main program exits, it presents an option to
keep the console window open (for looking at exceptions, etc.). If no action is
taken, the console window is closed after a timeout.

Currently, this is more a debugging tool than a Win32 launcher. It's helpful
nonetheless.