Add TYPE_POI_DB GeoDrawable type
TYPE_POI_DB is not actually drawable, but a database of points of interest.
It can be used to turn geographic names into coordinates.

Also, it is only implemented on the Python side. Very likely, this shouldn't be
here at all, but that's for later refactoring to rectify. Right now it solves
the problem of identifying this type of object for our map database / file list
on the Python side.

Quite a lot of the C++/Python interface is messy currently, and requires
cleanup. Unfortunately, I have not the time.