Add backend support for UTM/UPS coordinate systems
Add a class UTMUPS in ODM and export it via SIP.
Currently, conversions to/from LatLon are supported.

The bool field in UTMUPS required the suppression of MSVC warning C4800:
SIP generates code to cast an int directly to bool, which MSVC warns about.
Disable the warning, as we can't change SIPs code generation. It doesn't seem
tragic anyway.