Move map list handling to Python code
The only thing requiring RasterMapCollection on the C++ side was HeightFinder,
which needed it to find the best DHM to use.

Make HeightFinder derivable in Python, then we can locate the DHM there. Once
that is done, we can remove RasterMapCollection and friends.

LoadMap() was changed to return the constructed RasterMap directly,
and AlternateMapViews() was added to find more views given an already loaded
map.