Move map list handling from C++ to Python
To make this happen, we need to deal with HeightFinder: it needs to find an
appropriate DHM from the complete list of maps. We solve this by making
HeightFinder subclassable in Python. The map searching can then be relayed to
Python code via a virtual method call in C++. Sometimes, SIP is pretty neat.

With that in place, we can remove RasterMapCollection and
DefaultRasterMapCollection on the C++ side and handle everything in Python.

The fallout:
 * Change the way MainFrame instantiates the file list and height finder.
 * Remove failed previous attempts at doing the same thing
   (PythonRasterMapCollection).
 * Change GeoFilesCollection to track maps directly.
 * Clean up pymaplib/__init__.py, while already there.
 * Add aforementioned sub-classed HeightFinder
 * SIP changes galore, mostly deletions, changes to HeightFinder, a modified
   LoadMap() and a new AlternateMapViews().

A final noteworthy point:
  RasterMapShPtr must remain default- and move-constructable from SIP's point
  of view. Otherwise, moving it from Python back to C++ gives access
  violations. See the comment in rastermap.sip, look for AssignmentHelper in
  the SIP source.