Rework the internal file list
Stop storing GeoDrawables directly in our internal maplist/gpxlist/dblist --
ItemViews should be a real class, not a namedtuple. The main motivation for
this change are POI DBs, which belong in our file list, but are not drawable at
all.

So move all the file list handling to pymaplib/filelist.py, rename
GeoFilesCollection to FileList, and store FileListEntry instances in there.
FileListEntries also try to discourage direct use of file names by callers,
so we can include URLs or tokens in the future without breaking anything.

This had a large ripple effect on mapsevolved/frmMapManager.py, which was
cleaned up a bit in the process.