Fix some bugs in src_sip/std_types.sip
Refactor a bit, but many more problems remain. This file needs much more
serious refactoring, but I don't have the time right now.

Acute cause for this rework was hitting a segfault because of not checking
list entry type in the vector<wstring> "%ConvertToTypeCode". In some
circumstances we were trying to convert lists with arbitrary members to
vector<wstring>, and it blew up.