Add composite maps
Add composite maps: maps that consist of tiles of smaller maps files.
The usage scenario is tiling a large map (e.g. to avoid 4GB file size
restrictions), then using composite maps to represent them as one map in our
software.

Detailed list of changes:
 * Add class CompositeMap in OutdoorMapper/map_composite.{h,cpp}.
 * Add several string related utility functions (starts_with, url_{en,de}code).
 * Add logic to LoadMap to detect a "composite_map:" filename.
 * Add newly required functionality to class MapRegion.

A word about filenames/GetFname():
With this change we move away from GetFname() being an actual filename, towards
being more of a handle (or token) that allows a map to be created via
LoadMap(). Thus, users should not expect GetFname() to be a valid file!