Fix deleting of GPS or POI ErrorEntry items
We used to determine which list (maplist/gpxlist/dblist) to delete off based on
the entry_type of items. This failed when the file was missing or corrupted, in
which case we inserted an ErrorEntry, with entry_type == TYPE_ERROR.

Inject a new member ``ftype`` into the list items, which is set to either
"MAP", "GPX" or "DB". This tells us what we need to know on deletion. Take care
to also inject the original ftype for ErrorEntry items.