Make bitmap saving work for all maps
With some maps, a completely black bitmap was obtained when saving an area.

For these maps, DispOpenGL::RenderToBuffer() produced an PixelBuf with the
correct RGB values, but the alpha channel set to 0x00.
wx.Bitmap.CopyFromBuffer() with wx.BitmapBufferFormat_RGBA then interpreted
this as complete transparency.  Further, on Windows, the RGB pixels are
multiplied with the alpha channel, leading to the black image.

The best fix would be wx.BitmapBufferFormat_RGB32, but that swaps red and blue.
wx.BitmapBufferFormat_RGB doesn't work, as it loads 24 bits/pixel, not 32.

Fixing this on the OpenGL side is also difficult, as glReadPixels() is limited
in its output capabilities.

Thus, we fix it here by copying the read-only data to a writable bytearray and
manually setting all the alpha values to 255. That gives us a working image
in all cases, without a red-blue swap.