Fix map swithing without preserving map position
When switching from one map to another, mostly we try to preserve the current
position, if possible (i.e. if the current position is within the new map).
If that isn't possible, we want to show the center of the new map.

In practice, we computed the center for the old map instead. Whoops.