Implement GPS track display in Python
Update ODM to make drawing into PixelBuf's from Python easier.
Fix the SIP code so it's possible to implement GeoDrawables in Python.
Implement class GPSTrack in Python.

The eventual goal is to enable richer display of GPS tracks, with highlighting
of specific points, and tighter integration with the track analyzer.

This should also enable making a GPS track drawing tool, which would be a
valuable feature we are currently missing.