Implement drawing of new GPS tracks
Add the ability to draw new GPS tracks by clicking on the maps to create
waypoints. The resulting tracks are shown directly on the map, can be analyzed
with the GPS track editor, and can be saved once done.

This is a rudimentary implementation, with lots of features missing:
 * Editing tracks
 * Deleting the last drawn point
 * Tighter integration between the track display and the GPS analyzer (e.g.
   highlighting of selected points).

Nevertheless, it is already quite usable and works well in practice.