Allow access to GVGMap's internal objects
The intent is to make GVGMap more useful to for other use-cases, which need
more detailed knowledge of the GVG internals. The work of the accessor
functions could also be implemented in a user of the library, but it's harder,
since we'd need to figure out the best resolution again and retrieve the
mapinfo for it.

Since GVGMap has already done the heavy lifting, just let users access the
internal GVG/GMP objects.