Add gvg2geotiff executable
Add an executable to convert GVG maps into GeoTIFFs. The use-case is further
conversion into the OruxMaps OTK2 format for viewing georeferenced maps on
smartphones.

This makes it possible to use GVG encoded map data while on hiking trips. The
georeferenced nature makes it possible to show the current position directly
on the map.