Fix display problems with of stripped TIF images
LibTiff's TIFFRGBAImageGet has some limitations with stripped images
that we did not take into account. Work around those to get those
TIF's working properly.

While there, fix a range check in DoGetRegion() which gets rid of the
warnings sprewed to the console if the border of a map was displayed.