Remove HeightFinder, provide GetTerrainInfo()
Remove class HeightFinder, and provide the core functionality as a function
GetTerrainInfo().

HeightFinder relied on having a list of all maps at its disposal to find
the best DHM. This was unpractical ever since moving the maplist to Python.
We had to resort to a bit of a hack: subclass HeightFinder from Python
and do the map list handling there.

Remove this ugliness, instead provide GetTerrainInfo() which takes the map
to work on as an argument. Deciding which map to use can be done cleaner
on the Python side.