Move HeightFinder from C++ to Python
The C++ side was a hack, as it needed access to the map list, which is now
kept in Python. Remove this abomination and handle DHM selection fully on
the Python side.

This also fixes a few bugs with DHMs. Handling multiple, overlapping DHMs with
different resolutions should work fine now. The best resolution DHM is used
as long as it provides valid data, then progressively lower resolution maps
are used as fallback.