Rename BezierCoord to UnitSquareCoord
BezierCoords are not inherently tied to Bezier interpolation and can
also be used in other places. What distinguishes them is that the .x and
.y coordinates should be normalized (i.e. in [0, 1]), so BezierCoords
actually represent points inside the unit square.

Rename BezierCoords to reflect this.

We detect invalid values in the UnitSquareCoord constructor and raise an
std::runtime_error. We don't validate .x/.y anywhere else though, and
the fields are publically writeable. Should fix this sometime.