Saner bootstrapping of our build system
Do away with tools/init_venv.py. It was used to setup a venv, manually
install dependencies inside, to download UnxUtils and to download and
build the SIP distribution.

These tasks are now re-distributed as follows:

* We switch from Python 3.3 to 3.4 to get ensurepip.
* The venv setup is done by a new bootstrap.py, using pip install -r
  (also known as requirements.txt) to install the packages.
* UnxUtils and SIP are downloaded and built in ODM/libraries (via the
  invoke script there).

Some changes to tools/build.bat and tools/co-and-build.py were required
to make them deal with the new reality.