Allow building from already active venv
If a venv was already active and tools/init_shell.py was used to run
make, we didn't find link, cl, ...

This was caused by our PATH being reset to a previous value (without the
VS directories present) by activate.bat, if a venv was already active.

Now we explicitly deactivate any active venvs before re-initializing the
venv and the vsvars.