Add a CRT runtime library version checker
All the libraries we use in MapsEvolved must link against the same
version of the Microsoft Visual Studio C runtime library (CRT).
Otherwise we may get hard-to-reproduce failures and problems on
deployment.

In practice, we achieve this by compiling and linking all our libs with
the same version of Visual Studio and using the same compiler settings
(e.g. debug <-> release).

This patch adds tools/check_linked_libs.py, which analyzes if a list of
binaries all link against the same CRT version, and prints an error
afterwords. We now call this script at the end of our build process.

Further reading on this issue:
http://www.virtualdub.org/blog/pivot/entry.php?id=296
http://siomsystems.com/mixing-visual-studio-versions/
http://www.syndicateofideas.com/posts/fighting-the-msvcrt-dll-hell