Disable output buffering when calling sub-Pythons
In checkout_and_build and configure we run Python instances via
invoke.Runner.run(), which redirects output for logging purposes.

The sub-Pythons detect their stdout/stderr is not a tty and enable
aggressive output buffering, which results in weird ordering of log
messages when other programs (make, ...) writes to the console as well.

Set the PYTHONUNBUFFERED environment variable to remedy this.