Fix exceptions using TiffMap.LatLonToPixel()
We were getting exceptions when calling LatLonToPixel() on points
outside of a map (e.g. when switching to higher-resolution maps at the
current geographic position). This was tracked down to
BilinearInterpolator.inverse() creating a UnitSquareCoords outsie (0,1)
if the input LatLon is outside of the map.

This seems like a completely reasonable thing to do, so remove the value
checks from UnitSquareCoord.