Log map loading errors to the console
When loading of a map failes, print a short message to the console
describing the error. Should make it easier to see why maps fail to
load.

We need to turn the std::string into a std::wstring for output,
currently we hardcode UTF-8 as encoding. That's pretty shady. The real
fix is to bubble the error message up to the Python side so it can deal
with it (Python has more information about its output encoding and can
even decide to write to a logfile).

For now that's a quick fix to solve an immedate problem.