Test: Add a test file generator and some examples
create_testdata.py can currently be used to create test files with
geoinfo copied from currently installed GeoTIFFs.

tests/data/no-geoinfo.tif is the prototype TIF that the geoinfo is
implanted into. It is intentionally unsymmetric to avoid false positives
when testing if the image data can be read correctly.