Fix: Correct GPS scale when saving to bitmaps
When saving a bitmap with overlaid data (e.g. a GPS track), the overlay
was incorrectly sized compared to underlaying map.

This was tracked down to MapDisplayManager using a wrong display size
for SupportsDirectDrawing() == True maps, when called via PaintToBuffer.
It was retrieving the original (on-monitor) display size from the
display, instead of using the 'fake' (bitmap-output-) size supplied as
parameter.

Fix this by using the correct size.