Fix: Restore correct redraw on overlay change
Since the introduction of redraw-light, the display didn't update when
the overlay list was changed.

frmMain relied on a Refresh() call (i.e. InvalidateRect) to kick of a
redraw when the overlay list changed. With redraw-light, this didn't
set the need_full_repaint flag, so old display orders were used and the
updates weren't shown.

Export a ForceFullRepaint() method on MapDisplayManager to fix this.