C++: Move GetRegion calls to a background thread
Modify maps to opt-in to concurrent GetRegion() calls. For maps that
pass calls through to an underlying map that is trivial. For maps that
load data from a map file, mutexes were added.

This change eliminates the UI lockup that happened when large amounts of
map data were loaded when zooming out. Instead of locking up the UI
until everything is loaded, the image is now built sequentially while
the UI stays responsive.