Fix crash reading from stripped images
Concurrency testing revealed a reproducible crash when calling
Tiff::GetRegion() from multiple threads at the same time.

This turned out not to be caused by multi-threading, but only triggered
by it. The problem is reading stripped images where there are more rows
per strip than rows in the whole image (in other words, the image is not
even a whole strip high). If, for such an image, TiffRGBAImageGet() is
called with a requested height greater than the image height, an out of
bounds read happens. That this crashed only in a multi-threaded
scenario is likely due to chance (differing memory layout).

Fix this by capping the internally used strip size in GetRegion() to the
image height.

Technical background: TIFFStripSize() determines malloc() buffer size
and is capped to total_image_height. When the request size exceeds that
value, memory outside of the buffer is read. This gives either garbage
data (which is later discarded in GetRegion()), or a pagefault,
depending on the exact memory layout.