C++: Split MapView from MapDisplayManager
MapDisplayManager does a lot of things at once, and is thus reasonably
complicated. In particular, it conflated knowledge about the application
state with rendering.

Split out a class MapView that does all the rendering and shrink
MapDisplayManager to be more data-focused (and contain less logic).

This had ripple effects over a large part of the codebase, and the
result is not entirely clean yet, but it's a start. MapsEvolved runs on
the new foundations without apparent regressions.